+++ sigma/modules/moderation/punishments/softban.py
@@ -19,7 +19,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 import arrow
 import discord
 
-from sigma.core.utilities.data_processing import user_avatar
+from sigma.core.utilities.data_processing import user_avatar, get_broad_target
 from sigma.core.utilities.event_logging import log_event
 from sigma.core.utilities.generic_responses import denied, error
 from sigma.core.utilities.permission_processing import hierarchy_permit
@@ -58,8 +58,8 @@ async def softban(cmd, pld):
     :type pld: sigma.core.mechanics.payload.CommandPayload
     """
     if pld.msg.author.permissions_in(pld.msg.channel).ban_members:
-        if pld.msg.mentions:
-            target = pld.msg.mentions[0]
+        target = get_broad_target(pld)
+        if target:
             if cmd.bot.user.id != target.id:
                 if pld.msg.author.id != target.id:
                     above_hier = hierarchy_permit(pld.msg.author, target)